Completar información

+++ app/src/main/java/com/example/usersp/User.java
@@ -10,6 +10,6 @@ package com.cursosant.android.userssp
  * All my Courses(Only  on Udemy ):
  * https://www.udemy.com/user/alain-nicolas-tello/
  ****/
-data class  User  (val id: long, var name:String, var lastName: String,  var url:String){
+data class  User  (val id:int, var name:String, var lastName: String,  var url:String){
 
 }
